Support v0.2.0 and v0.3.0 (without atoms) mobiledoc versions
